柱状图 Style为STROKE 是 下边界的判断数据类型问题

+++ chart/src/main/java/cn/jingzhuan/lib/chart2/renderer/BarChartRenderer.java
@@ -158,7 +158,7 @@ public class BarChartRenderer extends AbstractDataRenderer<BarDataSet> {
                     top = top + strokeThickness;
                 }
                 if (barValue.getValueCount() > 1) bottom = calcHeight(barValue.getValues()[1], max, min);
-                if (bottom == mContentRect.height() && style == Paint.Style.STROKE) {
+                if ((int) bottom == mContentRect.height() && style == Paint.Style.STROKE) {
                     bottom = bottom - strokeThickness;
                 }
                 barValue.setCoordinate(x + width * 0.5f, top);
